renamed routers, controllers according to conventions. To make them easier to identify
